[REVIEW] rdar://99456859 (Low precision weight pickers do not maintain selected value when returning to question if .5 is selected, not whole number (#1278)
* by converting a double to int, we loose the deci

* unit test

+++ ResearchKit/Common/ORKWeightPicker.m
@@ -107,7 +107,7 @@ static const CGFloat PickerMinimumHeight = 34.0;
     
     if (_answerFormat.useMetricSystem) {
         if (_answerFormat.numericPrecision != ORKNumericPrecisionHigh) {
-            NSUInteger index = [_majorValues indexOfObject:@((NSInteger)((NSNumber *)answer).doubleValue)];
+            NSUInteger index = [_majorValues indexOfObject:@((double)((NSNumber *)answer).doubleValue)];
             if (index == NSNotFound) {
                 [self setAnswer:[self defaultAnswerValue]];
                 return;
